refactor cpp parsing module structure
