+++ src/lang/cpp/mod.rs
@@ -1,6 +1,14 @@
-use crate::ast::*;
 use crate::parse::AST;
 use crate::prophet::*;
+
+mod body;
+pub use body::*;
+
+mod expr;
+pub use expr::*;
+
+mod stmt;
+pub use stmt::*;
 
 pub fn merge_modules(modules: Vec<ModuleComponent>) -> Vec<ModuleComponent> {
     let mut merged = vec![];
@@ -189,7 +197,7 @@ fn transform_into_method(ast: &AST, module_name: &str, path: &str) -> Option<Met
         },
         None => (0, 0),
     };
-    let body = body.map_or_else(|| None, |body| Some(func_body(body)));
+    let body = body.map_or_else(|| None, |body| Some(body::func_body(body)));
 
     let method = MethodComponent {
         component: ComponentInfo {
@@ -510,637 +518,6 @@ fn field_is_abstract_method(field: &AST) -> bool {
     let eq = field.find_child_by_type(&["="]).is_some();
     let zero = field.find_child_by_value("0").is_some();
     virtual_specifier && eq && zero
-}
-
-// Takes in an AST with type field "compound_statement"
-fn func_body(body: &AST) -> Block {
-    let nodes = block_nodes(body);
-    Block::new(nodes)
-}
-
-fn block_nodes(compound_statement: &AST) -> Vec<Node> {
-    block_nodes_iter(&compound_statement.children)
-}
-
-fn block_nodes_iter(children: &[AST]) -> Vec<Node> {
-    children
-        .iter()
-        .map(|child| func_body_node(child))
-        .flat_map(|node| node)
-        .collect()
-}
-
-// Takes child of compound_statement
-fn func_body_node(node: &AST) -> Option<Node> {
-    match &*node.r#type {
-        "declaration" => {
-            let decl: Stmt = variable_declaration(node).into();
-            let decl = decl.into();
-            Some(decl)
-        }
-        "while_statement" => {
-            let cond = node
-                .find_child_by_type(&["condition_clause"])
-                .map(|cond| expression(cond))??;
-            let nodes = node
-                .find_child_by_type(&["compound_statement"])
-                .map(|block| block_nodes(block))?;
-            let while_stmt: Stmt = WhileStmt::new(cond, Block::new(nodes)).into();
-            Some(while_stmt.into())
-        }
-        "for_statement" => {
-            let for_stmt: Stmt = for_statement(node)?.into();
-            Some(for_stmt.into())
-        }
-        "for_range_loop" => {
-            let for_range_stmt: Stmt = for_range_statement(node)?.into();
-            Some(for_range_stmt.into())
-        }
-        "if_statement" => {
-            let if_stmt: Stmt = if_statement(node)?.into();
-            Some(if_stmt.into())
-        }
-        "switch_statement" => {
-            let switch_stmt: Stmt = switch_statement(node)?.into();
-            Some(switch_stmt.into())
-        }
-        "expression_statement" => {
-            let expr = node
-                .children
-                .iter()
-                .next()
-                .map_or_else(|| None, |node| expression(node))?;
-            let stmt: Stmt = expr.into();
-            Some(stmt.into())
-        }
-        "using_declaration" => {
-            let ident =
-                node.find_child_by_type(&["namespace_identifier", "scoped_namespace_identifier"])?;
-            let using = type_ident(ident);
-            let using: Stmt = ImportStmt::new(false, false, using).into();
-            Some(using.into())
-        }
-        "return_statement" => {
-            // If there isn't an expression and the 2nd child is of type ";",
-            // the expression function will return None anyways.
-            let expr = node
-                .children
-                .iter()
-                .nth(1)
-                .map(|expr| expression(expr))
-                .flatten();
-            let ret: Stmt = ReturnStmt::new(expr).into();
-            Some(ret.into())
-        }
-        "break_statement" => {
-            let brk: Stmt = BreakStmt::new().into();
-            Some(brk.into())
-        }
-        "continue_statment" => {
-            let cont: Stmt = ContinueStmt::new().into();
-            Some(cont.into())
-        }
-        "throw_statement" => {
-            let expr = expression(node.children.iter().nth(1)?);
-            let throw: Stmt = ThrowStmt::new(expr).into();
-            Some(throw.into())
-        }
-        "try_statement" => {
-            let mut children = node.children.iter();
-            let body = Block::new(block_nodes(children.nth(1)?));
-            let catch_bodies = children
-                .map(|catch| catch_statement(catch))
-                .filter_map(|catch| catch)
-                .collect();
-            let try_catch: Stmt = TryCatchStmt::new(body, catch_bodies).into();
-            Some(try_catch.into())
-        }
-        "compound_statement" => {
-            let nodes = block_nodes(node);
-            let block = Block::new(nodes);
-            Some(block.into())
-        }
-        // ...
-        _ => {
-            let expr: Stmt = expression(node)?.into();
-            Some(expr.into())
-        }
-    }
-}
-
-// Takes in node type "declaration"
-fn variable_declaration(node: &AST) -> DeclStmt {
-    let mut variable_type = node
-        .find_child_by_type(&[
-            "primitive_type",
-            "scoped_type_identifier",
-            "type_identifier",
-            "template_type",
-            "auto",
-        ])
-        .map_or_else(|| "".into(), |node| type_ident(node));
-
-    let init_declarator =
-        node.find_child_by_type(&["init_declarator", "function_declarator", "array_declarator"]);
-    match init_declarator {
-        Some(init_declarator) => match &*init_declarator.r#type {
-            "init_declarator" | "function_declarator" => {
-                variable_init_declaration(init_declarator, variable_type)
-            }
-            "array_declarator" => {
-                let rhs = expression(init_declarator).expect(&*format!(
-                    "Malformed array_declarator {:#?}",
-                    init_declarator
-                ));
-                DeclStmt::new(Some(variable_type), vec![rhs])
-            }
-            _ => unreachable!(),
-        },
-        None => {
-            let name = variable_ident(node, &mut variable_type).expect(&*format!(
-                "No variable name for declaration with no init {:#?}",
-                node,
-            ));
-            let ident = Ident::new(name);
-            DeclStmt::new(Some(variable_type), vec![ident.into()])
-        }
-    }
-}
-
-fn variable_init_declaration(init_declarator: &AST, mut variable_type: String) -> DeclStmt {
-    let name = variable_ident(init_declarator, &mut variable_type)
-        .expect("No identifier for init declarator");
-    let decl_type = init_declarator.find_child_by_type(&["=", "argument_list", "parameter_list"]);
-    let rhs = match decl_type {
-        Some(decl_type) => match &*decl_type.r#type {
-            "=" => {
-                let value = init_declarator.children.iter().next_back();
-                value.map_or_else(|| None, |value| expression(value))
-            }
-            "argument_list" | "parameter_list" => {
-                let argument_list = decl_type;
-                let args: Vec<Expr> = argument_list
-                    .children
-                    .iter()
-                    .map(|arg| expression(arg))
-                    .flat_map(|arg| arg)
-                    .collect();
-                let init = CallExpr::new(Box::new("new".to_string().into()), args).into();
-                Some(init)
-            }
-            _ => None,
-        },
-        None => None,
-    };
-    let ident = Ident::new(name);
-    let rhs = match rhs {
-        Some(rhs) => {
-            let init: Expr =
-                BinaryExpr::new(Box::new(ident.into()), Op::Equal, Box::new(rhs)).into();
-            vec![init]
-        }
-        None => vec![ident.into()],
-    };
-    DeclStmt::new(Some(variable_type), rhs)
-}
-
-fn expression(node: &AST) -> Option<Expr> {
-    match &*node.r#type {
-        "pointer_declarator" | "reference_declarator" | "parameter_declaration" => {
-            let mut ptr_symbol = String::new();
-            let name = variable_ident(node, &mut ptr_symbol)?;
-            let mut ident: Expr = match &*name {
-                "this" => Expr::Literal(name),
-                _ => Ident::new(name).into(),
-            };
-            ptr_symbol
-                .chars()
-                .map(|star| Op::from(&*star.to_string()))
-                .for_each(|star| ident = UnaryExpr::new(Box::new(ident.clone()), star).into());
-            Some(ident)
-        }
-        "pointer_expression" | "reference_expression" => {
-            let mut ptr_symbol = String::new();
-            let name = variable_ident_inner(node, &mut ptr_symbol)?;
-            let mut ident: Expr = match &*name {
-                "this" => Expr::Literal(name),
-                _ => Ident::new(name).into(),
-            };
-            ptr_symbol
-                .chars()
-                .map(|star| Op::from(&*star.to_string()))
-                .for_each(|star| ident = UnaryExpr::new(Box::new(ident.clone()), star).into());
-            Some(ident)
-        }
-        "identifier" | "field_identifier" | "type_identifier" => {
-            let name = node.value.clone();
-            let ident: Expr = Ident::new(name).into();
-            Some(ident)
-        }
-        "assignment_expression" | "binary_expression" => binary_expression(node), //| "field_expression"
-        "call_expression" => call_expression(node),
-        "field_expression" => {
-            let mut nodes = node.children.iter();
-            let lhs = expression(nodes.next()?)?;
-            let rhs = expression(nodes.last()?)?;
-            Some(DotExpr::new(Box::new(lhs), Box::new(rhs)).into())
-        }
-        "unary_expression" => {
-            let op = Op::from(&*node.children.iter().next()?.value);
-            let expr = expression(node.children.iter().last()?)?;
-            Some(UnaryExpr::new(Box::new(expr), op).into())
-        }
-        "parenthesized_expression" => {
-            let expr = node.children.iter().nth(1)?;
-            Some(ParenExpr::new(Box::new(expression(expr)?)).into())
-        }
-        "true" | "false" | "number_literal" | "this" | "auto" | "string_literal" => {
-            Some(node.value.clone().into())
-        }
-        "condition_clause" => {
-            let cond = node.children.iter().nth(1)?;
-            expression(cond)
-        }
-        // Handle scoped identifiers
-        "scoped_identifier" => {
-            let s = type_ident(node);
-            Some(Ident::new(s).into())
-        }
-        "update_expression" => {
-            // I need to potentially consider that an update expression may not be an IncDecExpr
-            let mut it = node.children.iter();
-            let first = it.next()?;
-            let second = it.next()?;
-            let update_expr_info = |node: &AST| match &*node.r#type {
-                "++" => (Some(true), Some(true)),
-                "--" => (Some(true), Some(false)),
-                _ => (None, None),
-            };
-            let (is_pre, is_inc) = update_expr_info(first);
-            match is_pre {
-                Some(_) => {
-                    let expr = expression(second)?;
-                    Some(IncDecExpr::new(is_pre?, is_inc?, Box::new(expr)).into())
-                }
-                None => {
-                    let expr = expression(first)?;
-                    let (is_pre, is_inc) = update_expr_info(second);
-                    Some(IncDecExpr::new(is_pre?, is_inc?, Box::new(expr)).into())
-                }
-            }
-        }
-        "array_declarator" | "subscript_expression" | "new_declarator" => {
-            // Identifier won't be found when the type is new_declarator.
-            let ident = expression(node.children.iter().nth(0)?)
-                .unwrap_or(Expr::Literal("Missing Ident".into()));
-            let ndx_expr = expression(node.children.iter().nth(2)?)?;
-            let ndx = IndexExpr::new(Box::new(ident), Box::new(ndx_expr));
-            Some(ndx.into())
-        }
-        "new_expression" => {
-            let r#type = expression(node.children.iter().nth(1)?)?;
-            let expr = expression(node.children.iter().last()?)?;
-            match expr {
-                Expr::IndexExpr(mut ndx) => {
-                    ndx.expr = Box::new(r#type);
-                    let init = CallExpr::new(Box::new("new".to_string().into()), vec![ndx.into()]);
-                    Some(init.into())
-                }
-                _ => None,
-            }
-        }
-        "delete_expression" => {
-            let expr = node.children.iter().find(|node| match &*node.r#type {
-                "delete" | "[" | "]" => false,
-                _ => true,
-            })?;
-            let expr = expression(expr)?;
-            let del = CallExpr::new(Box::new("delete".to_string().into()), vec![expr]);
-            Some(del.into())
-        }
-        "template_function" => {
-            let func_ident = Ident::new(type_ident(node));
-            Some(func_ident.into())
-        }
-        "initializer_list" => {
-            let exprs = node
-                .children
-                .iter()
-                .filter(|node| match &*node.r#type {
-                    "{" | "}" | "," => false,
-                    _ => true,
-                })
-                .map(|node| expression(node))
-                .flatten()
-                .collect();
-            Some(InitListExpr::new(exprs).into())
-        }
-        "sizeof_expression" => {
-            let expr = expression(node.children.iter().last()?)?;
-            let sizeof_call = CallExpr::new(Box::new("sizeof".to_string().into()), vec![expr]);
-            Some(sizeof_call.into())
-        }
-        "lambda_expression" => {
-            let parameter_list = node
-                .find_child_by_type(&["abstract_function_declarator"])?
-                .find_child_by_type(&["parameter_list"])?
-                .children
-                .iter()
-                .filter(|child| child.r#type == "parameter_declaration")
-                .map(|param_decl| variable_declaration(param_decl))
-                .collect();
-
-            let body = func_body(node.find_child_by_type(&["compound_statement"])?);
-
-            Some(LambdaExpr::new(parameter_list, body).into())
-        }
-        _ => None,
-    }
-}
-
-fn binary_expression(node: &AST) -> Option<Expr> {
-    let mut nodes = node.children.iter();
-    let lhs = expression(nodes.next()?)?;
-    let op = Op::from(&*nodes.next()?.value);
-    let rhs = expression(nodes.next()?)?;
-
-    // Check if the binary expression is a logging statement
-    // taking the form of (LOG(level) || cout) << ... << ...;
-    let is_log = match (&op, &lhs) {
-        // Identify the log call to know to create a LogExpr
-        (Op::BitShiftLeft, Expr::CallExpr(call)) => match &*call.name {
-            Expr::Ident(ident) => match &*ident.name.to_lowercase() {
-                "log" => true,
-                _ => false,
-            },
-            _ => false,
-        },
-        // LogExpr already created on left hand side, so append this expr
-        // to its arguments
-        (Op::BitShiftLeft, Expr::LogExpr(_)) => true,
-        // Printing to screen counts as a log
-        (Op::BitShiftLeft, Expr::Ident(ident)) => match &*ident.name {
-            "std::cout" | "cout" => true,
-            _ => false,
-        },
-        _ => false,
-    };
-
-    let expr = match lhs {
-        Expr::LogExpr(mut log) => {
-            log.args.push(rhs);
-            return Some(log.into());
-        }
-        _ => BinaryExpr::new(Box::new(lhs), op, Box::new(rhs)).into(),
-    };
-
-    if is_log {
-        let log = convert_binary_expr_to_log(expr)?;
-        Some(log.into())
-    } else {
-        Some(expr.into())
-    }
-}
-
-fn convert_binary_expr_to_log(cpp_log: BinaryExpr) -> Option<LogExpr> {
-    // Get the flattened children of the binary expression, with the left
-    // hand most child being the log call or cout identifier.
-    let mut children = flatten_binary_expr_children(cpp_log);
-    children.reverse();
-
-    // Get the log call expression
-    let log_call = match children.pop()? {
-        // LOG call identified
-        Expr::CallExpr(call) => call,
-        // Printed to screen, convert Ident to CallExpr
-        Expr::Ident(ident) => match &*ident.name {
-            "std::cout" | "cout" => CallExpr::new(
-                Box::new("".to_string().into()),
-                vec![Ident::new("console".into()).into()],
-            ),
-            _ => return None,
-        },
-        _ => return None,
-    };
-
-    children.reverse();
-
-    // Get the log level from the first argument.
-    let log_level = match &log_call.args.first()? {
-        Expr::Ident(ident) => LogLevel::from(&*ident.name),
-        _ => return None,
-    };
-
-    Some(LogExpr::new(log_level, children))
-}
-
-/// Convert all of the children of a binary expression into their left and right hand
-/// expressions, in the order they appear in the code.
-fn flatten_binary_expr_children(expr: BinaryExpr) -> Vec<Expr> {
-    match (*expr.lhs, *expr.rhs) {
-        (Expr::BinaryExpr(lhs), Expr::BinaryExpr(rhs)) => flatten_binary_expr_children(lhs)
-            .into_iter()
-            .chain(flatten_binary_expr_children(rhs))
-            .collect(),
-        (expr, Expr::BinaryExpr(rhs)) => vec![expr]
-            .into_iter()
-            .chain(flatten_binary_expr_children(rhs))
-            .collect(),
-        (Expr::BinaryExpr(lhs), expr) => flatten_binary_expr_children(lhs)
-            .into_iter()
-            .chain(vec![expr])
-            .collect(),
-        (lhs, rhs) => vec![lhs, rhs],
-    }
-}
-
-// Takes AST node type "call_expression"
-fn call_expression(node: &AST) -> Option<Expr> {
-    let mut nodes = node.children.iter();
-    // field_expression, identifier
-    let function_name = expression(nodes.next()?)?;
-    let argument_list = nodes.next_back()?;
-    let args: Vec<Expr> = argument_list
-        .children
-        .iter()
-        .map(|arg| expression(arg))
-        .flat_map(|arg| arg)
-        .collect();
-    Some(CallExpr::new(Box::new(function_name), args).into())
-}
-
-fn if_statement(if_stmt: &AST) -> Option<IfStmt> {
-    let cond = if_stmt
-        .find_child_by_type(&["condition_clause"])
-        .map(|cond| expression(cond))??;
-    let mut blocks = if_stmt
-        .children
-        .iter()
-        .filter(|node| &*node.r#type == "compound_statement")
-        .map(|block| Block::new(block_nodes(block)));
-    let body = match blocks.next() {
-        Some(block) => block,
-        None => {
-            let stmt = if_stmt.children.iter().last()?;
-            let stmt = func_body_node(stmt)?;
-            Block::new(vec![stmt])
-        }
-    };
-    // Check for else block, if else block, or no else block.
-    let else_body = match blocks.next() {
-        Some(else_body) => Some(else_body),
-        None => {
-            let else_if = if_stmt
-                .find_child_by_type(&["if_statement"])
-                .map(|if_stmt| if_statement(if_stmt))
-                .flatten()
-                .map(|if_stmt| {
-                    let if_stmt: Stmt = if_stmt.into();
-                    Block::new(vec![if_stmt.into()])
-                });
-            else_if
-        }
-    };
-    Some(IfStmt::new(cond, body, else_body))
-}
-
-fn switch_statement(switch_stmt: &AST) -> Option<SwitchStmt> {
-    let cond = switch_stmt
-        .find_child_by_type(&["condition_clause"])
-        .map(|cond| expression(cond))??;
-    let cases = switch_stmt
-        .find_child_by_type(&["compound_statement"])
-        .map(|switch_stmt| switch_stmt.children.iter())?
-        .map(|case| switch_case(case))
-        .flat_map(|case| case)
-        .collect();
-
-    let switch_stmt = SwitchStmt::new(cond, cases);
-    Some(switch_stmt)
-}
-
-fn switch_case(case_statement: &AST) -> Option<CaseStmt> {
-    let expr = case_statement.find_child_by_type(&["case", "default"])?;
-    // todo: add literals to expression function
-    let expr = match &*expr.r#type {
-        "case" => expression(case_statement.children.iter().nth(1)?),
-        "default" | _ => None,
-    };
-    let nodes = block_nodes_iter(&case_statement.children[3..]);
-    let block = Block::new(nodes);
-    let case = CaseStmt::new(expr, block);
-    Some(case.into())
-}
-
-fn for_statement(for_stmt: &AST) -> Option<ForStmt> {
-    let block = for_stmt.find_child_by_type(&["compound_statement"])?;
-    let block = func_body(block);
-
-    let mut init = None;
-    let mut cond = None;
-    let mut post = None;
-    let mut semicolons = 0u8;
-    for part in for_stmt
-        .children
-        .iter()
-        .filter(|child| match &*child.r#type {
-            "for" | "(" | ")" | "compound_statement" => false,
-            _ => true,
-        })
-    {
-        if &*part.r#type == ";" {
-            semicolons += 1;
-        } else {
-            match semicolons {
-                // Declarations need to be considered for the initialization. Regular BinExpr
-                // are treated as ExprStmt here.
-                0 => init = func_body_node(part),
-                1 => cond = expression(part),
-                2 => post = expression(part),
-                _ => {}
-            }
-        }
-    }
-
-    // Convert generic node to statement
-    let init = match init {
-        Some(Node::Stmt(stmt)) => Some(Box::new(stmt)),
-        _ => None,
-    };
-
-    let for_stmt = ForStmt::new(init, cond, post, block);
-    Some(for_stmt)
-}
-
-fn for_range_statement(for_range_loop: &AST) -> Option<ForRangeStmt> {
-    let block = for_range_loop.find_child_by_type(&["compound_statement"])?;
-    let block = func_body(block);
-
-    let mut i = 0;
-    let mut r#type = None;
-    let mut decl = None;
-    let mut iterator = None;
-    for part in for_range_loop
-        .children
-        .iter()
-        .filter(|child| match &*child.r#type {
-            "for" | "(" | ":" | ")" | "compound_statement" => false,
-            _ => true,
-        })
-    {
-        match i {
-            // Declarations need to be considered for the initialization. Regular BinExpr
-            // are treated as ExprStmt here.
-            0 => r#type = expression(part),
-            1 => decl = expression(part),
-            2 => iterator = expression(part),
-            _ => {}
-        }
-        i += 1;
-    }
-
-    // Convert generic node to statement
-    let mut r#type = match r#type {
-        Some(Expr::Literal(t)) => t,
-        Some(Expr::Ident(ident)) => ident.name,
-        _ => "".into(),
-    };
-    let mut is_unary = true;
-    while is_unary {
-        match decl {
-            Some(Expr::UnaryExpr(expr)) => {
-                r#type.push_str(match expr.op {
-                    Op::Star => "*",
-                    Op::And => "&",
-                    _ => "",
-                });
-                decl = Some(*expr.expr);
-                is_unary = true;
-            }
-            _ => is_unary = false,
-        }
-    }
-    let decl = DeclStmt::new(Some(r#type), vec![decl?]);
-    let for_range_stmt = ForRangeStmt::new(Box::new(decl.into()), iterator, block);
-    Some(for_range_stmt)
-}
-
-fn catch_statement(catch_clause: &AST) -> Option<CatchStmt> {
-    let params = catch_clause.find_child_by_type(&["parameter_list"])?;
-    let decl = params
-        .children
-        .iter()
-        .nth(1)
-        .map(|decl| match func_body_node(decl) {
-            Some(Node::Stmt(Stmt::DeclStmt(decl))) => Some(decl),
-            _ => None,
-        })
-        .flatten()
-        .unwrap_or(DeclStmt::new(None, vec![]));
-    let body = Block::new(block_nodes(
-        catch_clause.find_child_by_type(&["compound_statement"])?,
-    ));
-    Some(CatchStmt::new(decl, body))
 }
 
 #[cfg(test)]
@@ -1595,333 +972,5 @@ mod tests {
         };
         let destructor = transform_into_method(&destructor, "", "").unwrap();
         assert_eq!("~CastInfoServiceIf", destructor.method_name);
-    }
-
-    #[test]
-    fn return_stmt_test() {
-        let ast = AST {
-            children: vec![
-                AST {
-                    children: vec![],
-                    span: None,
-                    r#type: "return".to_string(),
-                    value: "return".to_string(),
-                },
-                AST {
-                    children: vec![
-                        AST {
-                            children: vec![],
-                            span: None,
-                            r#type: "!".to_string(),
-                            value: "!".to_string(),
-                        },
-                        AST {
-                            children: vec![
-                                AST {
-                                    children: vec![],
-                                    span: None,
-                                    r#type: "(".to_string(),
-                                    value: "(".to_string(),
-                                },
-                                AST {
-                                    children: vec![
-                                        AST {
-                                            children: vec![
-                                                AST {
-                                                    children: vec![],
-                                                    span: None,
-                                                    r#type: "*".to_string(),
-                                                    value: "*".to_string(),
-                                                },
-                                                AST {
-                                                    children: vec![],
-                                                    span: None,
-                                                    r#type: "this".to_string(),
-                                                    value: "this".to_string(),
-                                                },
-                                            ],
-                                            span: None,
-                                            r#type: "pointer_expression".to_string(),
-                                            value: "".to_string(),
-                                        },
-                                        AST {
-                                            children: vec![],
-                                            span: None,
-                                            r#type: "==".to_string(),
-                                            value: "==".to_string(),
-                                        },
-                                        AST {
-                                            children: vec![],
-                                            span: None,
-                                            r#type: "identifier".to_string(),
-                                            value: "rhs".to_string(),
-                                        },
-                                    ],
-                                    span: None,
-                                    r#type: "binary_expression".to_string(),
-                                    value: "".to_string(),
-                                },
-                                AST {
-                                    children: vec![],
-                                    span: None,
-                                    r#type: ")".to_string(),
-                                    value: ")".to_string(),
-                                },
-                            ],
-                            span: None,
-                            r#type: "parenthesized_expression".to_string(),
-                            value: "".to_string(),
-                        },
-                    ],
-                    span: None,
-                    r#type: "unary_expression".to_string(),
-                    value: "".to_string(),
-                },
-                AST {
-                    children: vec![],
-                    span: None,
-                    r#type: ";".to_string(),
-                    value: ";".to_string(),
-                },
-            ],
-            span: None,
-            r#type: "return_statement".to_string(),
-            value: "".to_string(),
-        };
-
-        let expected: Stmt = ReturnStmt::new(Some(
-            UnaryExpr::new(
-                Box::new(
-                    ParenExpr::new(Box::new(
-                        BinaryExpr::new(
-                            Box::new(
-                                UnaryExpr::new(Box::new(Expr::Literal("this".into())), Op::Star)
-                                    .into(),
-                            ),
-                            Op::EqualEqual,
-                            Box::new(Ident::new("rhs".into()).into()),
-                        )
-                        .into(),
-                    ))
-                    .into(),
-                ),
-                Op::ExclamationPoint,
-            )
-            .into(),
-        ))
-        .into();
-        let expected: Node = expected.into();
-
-        let actual = func_body_node(&ast).unwrap();
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn init_declaration_test() {
-        let ast = AST {
-            children: vec![
-                AST {
-                    children: vec![],
-                    span: None,
-                    r#type: "primitive_type".to_string(),
-                    value: "uint32_t".to_string(),
-                },
-                AST {
-                    children: vec![
-                        AST {
-                            children: vec![],
-                            span: None,
-                            r#type: "identifier".to_string(),
-                            value: "xfer".to_string(),
-                        },
-                        AST {
-                            children: vec![],
-                            span: None,
-                            r#type: "=".to_string(),
-                            value: "=".to_string(),
-                        },
-                        AST {
-                            children: vec![],
-                            span: None,
-                            r#type: "number_literal".to_string(),
-                            value: "0".to_string(),
-                        },
-                    ],
-                    span: None,
-                    r#type: "init_declarator".to_string(),
-                    value: "".to_string(),
-                },
-                AST {
-                    children: vec![],
-                    span: None,
-                    r#type: ";".to_string(),
-                    value: ";".to_string(),
-                },
-            ],
-            span: None,
-            r#type: "declaration".to_string(),
-            value: "".to_string(),
-        };
-
-        let expected: Stmt = DeclStmt::new(
-            Some("uint32_t".into()),
-            vec![BinaryExpr::new(
-                Box::new(Ident::new("xfer".into()).into()),
-                Op::Equal,
-                Box::new(Expr::Literal("0".into())),
-            )
-            .into()],
-        )
-        .into();
-        let expected: Node = expected.into();
-
-        let actual = func_body_node(&ast).unwrap();
-
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn empty_for_loop_test() {
-        let ast = AST {
-            children: vec![
-                AST {
-                    children: vec![],
-                    span: None,
-                    r#type: "for".to_string(),
-                    value: "for".to_string(),
-                },
-                AST {
-                    children: vec![],
-                    span: None,
-                    r#type: "(".to_string(),
-                    value: ")".to_string(),
-                },
-                AST {
-                    children: vec![
-                        AST {
-                            children: vec![],
-                            span: None,
-                            r#type: "identifier".to_string(),
-                            value: "_i284".to_string(),
-                        },
-                        AST {
-                            children: vec![],
-                            span: None,
-                            r#type: "=".to_string(),
-                            value: "=".to_string(),
-                        },
-                        AST {
-                            children: vec![],
-                            span: None,
-                            r#type: "number_literal".to_string(),
-                            value: "0".to_string(),
-                        },
-                    ],
-                    span: None,
-                    r#type: "assignment_expression".to_string(),
-                    value: "".to_string(),
-                },
-                AST {
-                    children: vec![],
-                    span: None,
-                    r#type: ";".to_string(),
-                    value: ";".to_string(),
-                },
-                AST {
-                    children: vec![
-                        AST {
-                            children: vec![],
-                            span: None,
-                            r#type: "identifier".to_string(),
-                            value: "_i284".to_string(),
-                        },
-                        AST {
-                            children: vec![],
-                            span: None,
-                            r#type: "<".to_string(),
-                            value: "<".to_string(),
-                        },
-                        AST {
-                            children: vec![],
-                            span: None,
-                            r#type: "identifier".to_string(),
-                            value: "_size280".to_string(),
-                        },
-                    ],
-                    span: None,
-                    r#type: "binary_expression".to_string(),
-                    value: "".to_string(),
-                },
-                AST {
-                    children: vec![],
-                    span: None,
-                    r#type: ";".to_string(),
-                    value: ";".to_string(),
-                },
-                AST {
-                    children: vec![
-                        AST {
-                            children: vec![],
-                            span: None,
-                            r#type: "++".to_string(),
-                            value: "++".to_string(),
-                        },
-                        AST {
-                            children: vec![],
-                            span: None,
-                            r#type: "identifier".to_string(),
-                            value: "_i284".to_string(),
-                        },
-                    ],
-                    span: None,
-                    r#type: "update_expression".to_string(),
-                    value: "".to_string(),
-                },
-                AST {
-                    children: vec![],
-                    span: None,
-                    r#type: ")".to_string(),
-                    value: ")".to_string(),
-                },
-                AST {
-                    children: vec![],
-                    span: None,
-                    r#type: "compound_statement".to_string(),
-                    value: "".to_string(),
-                },
-            ],
-            span: None,
-            r#type: "for_statement".to_string(),
-            value: "".to_string(),
-        };
-
-        let init: Expr = BinaryExpr::new(
-            Box::new(Ident::new("_i284".into()).into()),
-            Op::Equal,
-            Box::new(Expr::Literal("0".into())),
-        )
-        .into();
-        let init: Stmt = init.into();
-
-        let expected: Stmt = ForStmt::new(
-            Some(Box::new(init)),
-            Some(
-                BinaryExpr::new(
-                    Box::new(Ident::new("_i284".into()).into()),
-                    Op::LessThan,
-                    Box::new(Ident::new("_size280".into()).into()),
-                )
-                .into(),
-            ),
-            Some(IncDecExpr::new(true, true, Box::new(Ident::new("_i284".into()).into())).into()),
-            Block::new(vec![]).into(),
-        )
-        .into();
-        let expected: Node = expected.into();
-
-        let actual = func_body_node(&ast).unwrap();
-
-        assert_eq!(expected, actual);
     }
 }
